requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Local LLM Integration
ollama>=0.1.6

# LLM Orchestration & RAG
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Web Search
tavily-python>=0.3.0

# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.1.0
html2text==2020.1.16

# Document Generation - PDF
reportlab==4.0.9
weasyprint==60.2

# Document Generation - DOCX
python-docx==1.1.0

# Data Processing
numpy>=1.26.3
pandas>=2.1.4

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.3
tqdm==4.66.1

# Date/Time
python-dateutil>=2.8.2
